debut message pour demander donner d'un projet

+++ src/client/clientnetwork.hpp
@@ -25,7 +25,7 @@ public:
     void getUsersProjects(long long id);
     void getProjectList(); 
     
-    void getProjectData(int project_id);
+    void getProjectData(long long project_id);
     void delProject();
     void renameProject();
     void createProjectCode();
